Blog2Video/requirements.txt
keybert
gttS
imageio==2.4.1
imageio-ffmpeg
moviepy
bing-image-downloader
pillow